+++ lib/smart_pagination/widgets/smart_pagination_page_view.dart
@@ -55,6 +55,8 @@ class SmartPaginationPageView<T> extends SmartPagination<T> {
     super.loadMoreErrorBuilder,
     super.loadMoreNoMoreItemsBuilder,
     super.invisibleItemsThreshold,
+    super.canRefresh,
+    super.onRefresh,
     super.refreshListener,
     super.filterListeners,
     super.onInsertionCallback,
@@ -103,6 +105,8 @@ class SmartPaginationPageView<T> extends SmartPagination<T> {
     super.loadMoreErrorBuilder,
     super.loadMoreNoMoreItemsBuilder,
     super.invisibleItemsThreshold,
+    super.canRefresh,
+    super.onRefresh,
     super.refreshListener,
     super.filterListeners,
   }) : super.pageViewWithCubit();