+++ lib/smart_pagination/widgets/smart_pagination_reorderable_list_view.dart
@@ -63,6 +63,8 @@ class SmartPaginationReorderableListView<T> extends SmartPagination<T> {
     super.loadMoreErrorBuilder,
     super.loadMoreNoMoreItemsBuilder,
     super.invisibleItemsThreshold,
+    super.canRefresh,
+    super.onRefresh,
     super.refreshListener,
     super.filterListeners,
     super.onInsertionCallback,
@@ -111,6 +113,8 @@ class SmartPaginationReorderableListView<T> extends SmartPagination<T> {
     super.loadMoreErrorBuilder,
     super.loadMoreNoMoreItemsBuilder,
     super.invisibleItemsThreshold,
+    super.canRefresh,
+    super.onRefresh,
     super.refreshListener,
     super.filterListeners,
   }) : super.reorderableListViewWithCubit();