+++ lib/smart_pagination/widgets/smart_pagination_row.dart
@@ -55,6 +55,8 @@ class SmartPaginationRow<T> extends SmartPagination<T> {
     super.loadMoreErrorBuilder,
     super.loadMoreNoMoreItemsBuilder,
     super.invisibleItemsThreshold,
+    super.canRefresh,
+    super.onRefresh,
     super.refreshListener,
     super.filterListeners,
     super.onInsertionCallback,
@@ -99,6 +101,8 @@ class SmartPaginationRow<T> extends SmartPagination<T> {
     super.loadMoreErrorBuilder,
     super.loadMoreNoMoreItemsBuilder,
     super.invisibleItemsThreshold,
+    super.canRefresh,
+    super.onRefresh,
     super.refreshListener,
     super.filterListeners,
   }) : super.rowWithCubit();